Convert QDomDocument to QXmlStreamWriter for saving XML files
Because QDomDocument does not preserve the ordering of attributes,
the conversion to QString gives a random order of XML nodes attributes.
This is an issue if you consider using a VCS (such as git) to manage basket
directory.

The proper way is to write XML files using QXmlStreamWriter, which is also much
more efficient.