Remove the global KAboutData object, fix translations
The global KAboutData was initialized before Application, which
means before QApplication and the setting of the domain,
and then the content of the KAboutData object was not translated.
This change removes the global KAboutData object
but makes sure that data are not duplicated.

It also sets the translation domain immediately after
the constuction of the QApplication instance,
so that, Without the global object, the translations are
correctly applied to the entire program.