Fix virtual function call in FileContent
the virtual linkLook() was called in constructor which is fixed
trivially.
However, SoundContent inherits FileContent in which SoundContent
intends to override the linkLook() function, but is not available
at during construction.
I had to copy the setFileName method to SoundContent and applied
the explicit function call. However, this is very inconvinent,
and disputable.