Fix typo in Welcome basket (en_US)
Fixes the typo in the sentence:
"This multi-purpose note-taking application can helps you to:"
To:
"This multi-purpose note-taking application can help you to:"

This baskets are loaded on first use, meant as an example and
kind of tutorial. They have a custom binary format which enclose
a preview `.png`and a `tar.gz`.

BUG: 287729
FIXED-IN: ¯\_(ツ)_/¯