flatpak: convert the manifest to json
Same content, plus the new separate file for the license.

Leaving aside the personal ideas of the author of this patch
about which one between YAML and JSON is more readable, all the
other flatpak manifests on kde.org are in JSON format.
Consistency is important enough to change it.

Signed-off-by: Luigi Toscano <luigi.toscano@tiscali.it>